+++ FolioClasses/FolioClasses.csproj
@@ -46,7 +46,7 @@
     <Compile Include="AuthorManagement\ClsAuthor.cs" />
     <Compile Include="ClsDataConnection.cs" />
     <Compile Include="clsOrder.cs" />
-    <Compile Include="UserManagement\ClsCustomerUser.cs" />
+    <Compile Include="UserManagement\clsCustomerUser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PublisherManagement\ClsPublisher.cs" />
   </ItemGroup>
